test: 테이블 테스트 분리
